IIssueFixes64_68_72_76_161_196_269 (#78)
Issue161: ResHvac.xsd corrected spelling of several Rheem HPWH Values
and removing duplicates with the typo in multiple data types that
included these values
Issue 64: LMCI/LMCV-MCH-23e: E01 pseudo code added to cover MCH01b as
parent doc
Issue 68: LMCI-MCH-01c C04 added missing DuctedMultiSplitHP to list in
pseudocode
Issue 72: Removed fields not currently supported in the software by
setting minOccurs and maxOccurs to zero which should not require any
changes in rendering code. This type of changes was made to LMCI-MCH01d:
O11, LMCI/LMCV-MCH-22 and 23 A12. The other change for this issue is
marking references to A12 as NOT IMPLEMENTED in the fields that assigns
the variant to use. For the MCH-22's this is B01After. For MCH-23's it
is C04After
Issue 76: LMCI/LMCV-MCH-22c D03 pseudo code add logic for MCH01b or
MCH01c parent
Issues 196 and 269: LMCI/LMCV-PLB01/21 C03 and LMCI/LMCV-PLB02/22 A04
and B04 added choice with WaterHeaterType

---------

Co-authored-by: Actions Bot <scripts-bot@rasentsolutions.com>